requirements.txt
asgiref==3.5.1
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
ci-info==0.2.0
circlify==0.14.0
click==8.1.3
colorama==0.4.4
configobj==5.0.6
configparser==5.2.0
cycler==0.11.0
decorator==5.1.1
Django==4.0
et-xmlfile==1.1.0
etelemetry==0.3.0
executing==0.8.3
filelock==3.6.0
fitz==0.0.1.dev2
fonttools==4.33.3
future==0.18.2
httplib2==0.20.4
idna==3.3
ipython==8.3.0
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.2
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mpld3==0.5.7
nibabel==3.2.2
nipype==1.8.0
nltk==3.7
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathlib==1.0.1
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
prov==2.0.0
pure-eval==0.2.2
pydot==1.4.2
Pygments==2.12.0
PyMuPDF==1.19.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pyxnat==1.4
rdflib==6.1.1
regex==2022.4.24
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tqdm==4.64.0
traitlets==5.2.0
traits==6.3.2
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5

PyPDF2
pdfplumber
wordcloud
seaborn
sklearn
tweepy
xlsxwriter
spacy